LayoutTests/css3/flexbox/new-alignment-values-invalid-if-grid-not-enabled-expected.txt

Test to verify that the new alignment values are parsed as invalid if Grid Layout is disabled and in any case they do not cause a crash because assertions in flexbox layout code.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



New alignment values should be invalid when grid layout is disabled

Testing Self-Alignment values.
PASS alignSelf is 'flex-start'
PASS alignSelf is 'flex-start'
PASS alignSelf is 'flex-start'
PASS alignSelf is 'flex-start'
PASS alignSelf is 'flex-start'
PASS alignSelf is 'flex-start'

Testing Default-Alignment values.
PASS alignItems is 'flex-end'
PASS alignSelf is 'flex-end'
PASS alignItems is 'flex-end'
PASS alignSelf is 'flex-end'
PASS alignItems is 'flex-end'
PASS alignSelf is 'flex-end'
PASS alignItems is 'flex-end'
PASS alignSelf is 'flex-end'
PASS alignItems is 'flex-end'
PASS alignSelf is 'flex-end'
PASS alignItems is 'flex-end'
PASS alignSelf is 'flex-end'

Even when grid layout is enabled, new values should not violate assertions in FlexibleBox layout logic.

Testing Self-Alignment values.
PASS alignSelf is 'start unsafe'
PASS alignSelf is 'start'
PASS alignSelf is 'end'
PASS alignSelf is 'flex-start safe'
PASS alignSelf is 'self-start'
PASS alignSelf is 'self-end'

Testing Default-Alignment values.
PASS alignItems is 'start unsafe'
PASS alignSelf is 'start unsafe'
PASS alignItems is 'start'
PASS alignSelf is 'start'
PASS alignItems is 'end'
PASS alignSelf is 'end'
PASS alignItems is 'flex-start safe'
PASS alignSelf is 'flex-start safe'
PASS alignItems is 'self-start'
PASS alignSelf is 'self-start'
PASS alignItems is 'self-end'
PASS alignSelf is 'self-end'
PASS successfullyParsed is true

TEST COMPLETE
